[wip] feat: 灵脉系统
